fix(term): update xterm from 3.14.5 to 5.4.0
Xterm v3 contained a redraw bug when you clicked on the canvas, then
clicked outside of the canvas, then execute `term.write('foo\n')`
multiple times, this would run in a redraw bug where it still shows the
cursor in places where it was before, but is not anymore. The new v5
fixes this and also uses HTML elements rather than a canvas element.