fix: adjust styling such that the term is always within viewport
When GoldenLayout is initialized, it creates a new container within the
layout-container, which made the container slightly higher than the
viewport, because the container gets pushed down by the navbar. Forcing
the height of the goldenlayout container to be the parent's height fixes
this issue, as goldenlayout doesn't take into account the fact that
there is a navbar outside of it's container.